apply-externals: optimize grep stage
This optimization caches the output of the grep pass (used to reduce the
listing of the whole repository returned by ls-tree to just the paths of
interest). ls-tree provides this functionality, but it seems to be slower
to repeatedly invoke ls-tree (even with caching its output) than to
invoke ls-tree once and run multiple greps over it.

This optimization provides ~130% speed improvement in my case.